remove inference sample of uie
