--- model_zoo/uie/deploy/python/requirements_cpu.txt
@@ -1,2 +1,0 @@
-onnx
-onnxruntime
